Former-commit-id: 4ad86d7db7e5a563aa3039a65dab20bff16cf044

+++ openquake/commands/purge.py
@@ -20,9 +20,10 @@ from __future__ import print_function
 import os
 import re
 import getpass
+import shutil
 from openquake.baselib import sap
 from openquake.commonlib import datastore
-from openquake.engine.logs import dbcmd
+from openquake.commonlib.logs import dbcmd
 
 
 def purge_one(calc_id, user):
@@ -39,22 +40,29 @@ def purge_one(calc_id, user):
 
 
 # used in the reset command
-def purge_all(user=None):
+def purge_all(user=None, fast=False):
     """
     Remove all calculations of the given user
     """
     user = user or getpass.getuser()
-    for fname in os.listdir(datastore.DATADIR):
-        mo = re.match('calc_(\d+)\.hdf5', fname)
-        if mo is not None:
-            calc_id = int(mo.group(1))
-            purge_one(calc_id, user)
+    datadir = datastore.DATADIR
+    if os.path.exists(datadir):
+        if fast:
+            shutil.rmtree(datadir)
+            print('Removed %s' % datadir)
+        else:
+            for fname in os.listdir(datadir):
+                mo = re.match('calc_(\d+)\.hdf5', fname)
+                if mo is not None:
+                    calc_id = int(mo.group(1))
+                    purge_one(calc_id, user)
 
 
 @sap.Script
 def purge(calc_id):
     """
-    Remove the given calculation. If calc_id is 0, remove all calculations.
+    Remove the given calculation. If you want to remove all calculations,
+    use oq reset.
     """
     if calc_id < 0:
         try:
